+++ PantherGUI/control.cpp
@@ -22,7 +22,7 @@ void Control::Draw(PGRendererHandle handle, PGIRect* rectangle) {
 void Control::PeriodicRender(void) {
 }
 
-void Control::MouseWheel(int x, int y, int distance, PGModifier modifier) {
+void Control::MouseWheel(int x, int y, double distance, PGModifier modifier) {
 
 }
